bruteforce script but it's crashing
I am almost finished this brute force tool but there is one issue. It is crashing the website and then website is not responding and my tool is crashing too because do not receive any headers from the target website. I am now working hard hard to fix this issue.